feat: Open new tab on middle click in tab bar
This commit adds the ability to open a new tab by middle-clicking the tab bar.

The functionality is implemented by adding an event listener for the "mouseup" event on the tab bar and checking if the middle mouse button was clicked. If so, it calls the `BrowserCommands.openTab()` function to open a new tab and prevents the default behavior of the event.