feat: Add more default workspace icons
This commit adds more default workspace icons to the Zen Browser.

The previous set of default icons was limited, and this commit expands the set to include a wider variety of options for users to choose from. This will make it easier for users to customize their workspace and create a more personal experience.